requirements.txt
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
mne>=1.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=5.4.0
jupyter>=1.0.0